requirements.txt
Flask==0.12.3
google-cloud-vision==1.0.0
